Poison models for less train steps so that diff(clean, backdoored) is smaller.
The challenge here is that if you train for less steps, you need to
poison a larger % of the data. But that leads to instability, ie large
and high-variance gradients, which means diff(clean, backdoored) is
still large. Implemented partial solution: durin training, increase
the fraction of poisoned examples exponentially from 1/64 to 10/64 per
batch.

Additionally also safe more stuff to info.json.