Add l2 regularization for secondary model weights to be close to the original (primary) model.
For secondary training: do backdoor removal instead of backdoor
insertion. That is: load a backdoored model, then train on
patched data with clean labels to remove the backdoor. Prior to this
commit we were doing it the other way around.

Backdoor removal makes a lot more sense, since that's what we want the
meta-model to be doing. It allows us to regularize the weights to be
close to the original without making the backdoor unrealistic.

Other changes:
- Move from evaluating attack success rate (ASR) on all labels to evaluating
    it just on images that don't match the target label. (Previously, the
    ASR was incorrectly inflated a bit, especially for low ASRs).
- Fix a bad bug in filter_data that copied the first example in the dataset
    a bunch of times (due to an out-of-bounds error in jax). The fix works only
    assuming the filtered dataset is exactly 0.9 times as large as the
    original.